// src/server/api/models/administration/models/UserPasswordChangeRequest.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * A JSON blob representing a user password change request.
 */
export type UserPasswordChangeRequest = {
    /**
     * The user name
     */
    username: string;
    /**
     * The password
     */
    password: string;
};